BHI firmware can be uploaded directly
Load the firmware binary file directly into the main.cpp and then write it to the SPI Flash attached to nrf52832.

In addition, the CRC of the original content and the flash content are checked to see if they match.

The firmware and then the CRC for the firmware were written to the SPI flash